+++ out/functions/bundle.js
@@ -12,6 +12,7 @@ exports.default = {
     openQuickPick: openQuickPick_1.default,
     createInputBox: createInputBox_1.default,
     isInputable: isInputable_1.default,
-    changeInputVariable: changeInputvariable_1.default,
+    non_optional_changer: changeInputvariable_1.non_optional_changer,
+    optional_changer: changeInputvariable_1.optional_changer,
 };
 //# sourceMappingURL=bundle.js.map