+++ src/commands/openList.ts
@@ -18,7 +18,7 @@ export default async function OpenList(): Promise<void> {
     let booleanQuickPick = functions.openQuickPick(
         "Create custom_commands file?",
         items,
-        onSelect
+        onSelect,
     );
     let ifError = (message: string): boolean => {
         booleanQuickPick.show();
@@ -33,7 +33,7 @@ export default async function OpenList(): Promise<void> {
         async function getSyncData() {
             fileData = await functions.readFile(
                 vars().ccName + vars().ccExtension,
-                ""
+                "",
             );
             splittedData = fileData.split("\n");
         }
@@ -67,7 +67,7 @@ export default async function OpenList(): Promise<void> {
             function clearTerminal() {
                 let terminalName: string = vars().terminalPath.split("\\");
                 terminalName = terminalName[terminalName.length - 1].split(
-                    "."
+                    ".",
                 )[0];
                 return terminalName;
             }
@@ -78,26 +78,84 @@ export default async function OpenList(): Promise<void> {
                 clearCommand = "clear";
             }
             let line_text: string = item[0].label;
-            if (functions.isInputable(line_text)) {
-                await functions
-                    .changeInputVariable(line_text)
-                    .then((res: string) => {
-                        vscode.commands.executeCommand(
-                            "workbench.action.focusPanel"
-                        );
-                        let newTerminal: vscode.Terminal = vars().createTerminal(
-                            vars().customTerminalName + " - 1",
-                            vars().terminalPath
-                        );
-                        newTerminal.show();
-                        newTerminal.sendText(clearCommand);
-                        newTerminal.sendText(variableTransformer(res));
-                    });
+            if (functions.isInputable(line_text).result == true) {
+                let types_input = await functions.isInputable(line_text).types;
+                if (
+                    types_input.non_optional === true &&
+                    types_input.optional === true
+                ) {
+                    await functions
+                        .non_optional_changer(line_text)
+                        .then(async (res) => {
+                            if (res.cancelled == true) {
+                                return;
+                            }
+
+                            await functions
+                                .optional_changer(res.line)
+                                .then((new_res) => {
+                                    vscode.commands.executeCommand(
+                                        "workbench.action.focusPanel",
+                                    );
+                                    let newTerminal: vscode.Terminal = vars().createTerminal(
+                                        vars().customTerminalName + " - 1",
+                                        vars().terminalPath,
+                                    );
+                                    newTerminal.show();
+                                    newTerminal.sendText(clearCommand);
+                                    newTerminal.sendText(
+                                        variableTransformer(new_res.line),
+                                    );
+                                });
+                        });
+                } else if (
+                    types_input.non_optional === true &&
+                    types_input.optional === false
+                ) {
+                    await functions
+                        .optional_changer(line_text)
+                        .then((res: { cancelled: boolean; line: string }) => {
+                            if (res.cancelled === true) {
+                                return;
+                            } else if (res.cancelled === false) {
+                                vscode.commands.executeCommand(
+                                    "workbench.action.focusPanel",
+                                );
+                                let newTerminal: vscode.Terminal = vars().createTerminal(
+                                    vars().customTerminalName + " - 1",
+                                    vars().terminalPath,
+                                );
+                                newTerminal.show();
+                                newTerminal.sendText(clearCommand);
+                                newTerminal.sendText(
+                                    variableTransformer(res.line),
+                                );
+                            }
+                        });
+                } else if (
+                    types_input.non_optional === false &&
+                    types_input.optional === true
+                ) {
+                    await functions
+                        .optional_changer(line_text)
+                        .then((res: { line: string }) => {
+                            vscode.commands.executeCommand(
+                                "workbench.action.focusPanel",
+                            );
+                            let newTerminal: vscode.Terminal = vars().createTerminal(
+                                vars().customTerminalName + " - 1",
+                                vars().terminalPath,
+                            );
+                            newTerminal.show();
+                            newTerminal.sendText(clearCommand);
+                            newTerminal.sendText(variableTransformer(res.line));
+                        });
+                }
             } else {
                 vscode.commands.executeCommand("workbench.action.focusPanel");
                 let newTerminal: vscode.Terminal = vars().createTerminal(
                     vars().customTerminalName + " - 1",
-                    vars().terminalPath
+                    vars().terminalPath,
                 );
                 newTerminal.show();
                 newTerminal.sendText(clearCommand);
@@ -108,7 +166,7 @@ export default async function OpenList(): Promise<void> {
         let ListQuickPick = functions.openQuickPick(
             "Select Command From List",
             lines,
-            onChoose
+            onChoose,
         );
         ListQuickPick.show();
     }