+++ src/functions/bundle.ts
@@ -3,7 +3,7 @@ import readFile from "./readFile";
 import openQuickPick from "./openQuickPick";
 import createInputBox from "./createInputBox";
 import isInputable from "./isInputable";
-import changeInputVariable from "./changeInputvariable";
+import { non_optional_changer, optional_changer } from "./changeInputvariable";
 
 export default {
     createFile,
@@ -11,5 +11,6 @@ export default {
     openQuickPick,
     createInputBox,
     isInputable,
-    changeInputVariable,
+    non_optional_changer,
+    optional_changer,
 };